+++ src/config/common.js
@@ -4,8 +4,8 @@ const includeRules = require('../lib/include-rules');
 
 module.exports = (configs, options) => ({
 	parserOptions: {
-		ecmaVersion: 6,
-		sourceType: 'module',
+		ecmaVersion: 11,
+		sourceType: 'script',
 		ecmaFeatures: {
 			globalReturn: true,
 			implicitStrict: false,
@@ -14,9 +14,12 @@ module.exports = (configs, options) => ({
 		},
 	},
 	plugins: [
+		'eslint-comments',
 		'filenames',
+		'import',
 		'promise',
 		'security',
+		'unicorn',
 	],
 	env: {
 		'shared-node-browser': true,
@@ -24,13 +27,19 @@ module.exports = (configs, options) => ({
 	},
 	rules: {
 		...includeRules('best-practices', configs, options),
-		...includeRules('ecmascript-6', configs, options),
+		...includeRules('ecmascript-2015', configs, options),
+		...includeRules('ecmascript-2017', configs, options),
+		...includeRules('ecmascript-2018', configs, options),
 		...includeRules('possible-errors', configs, options),
 		...includeRules('strict-mode', configs, options),
 		...includeRules('stylistic-issues', configs, options),
 		...includeRules('variables', configs, options),
+		...includeRules('plugins/eslint-comments/best-practices', configs, options),
+		...includeRules('plugins/eslint-comments/stylistic-issues', configs, options),
 		...includeRules('plugins/filenames', configs, options),
+		...includeRules('plugins/import', configs, options),
 		...includeRules('plugins/promise', configs, options),
 		...includeRules('plugins/security', configs, options),
+		...includeRules('plugins/unicorn', configs, options),
 	},
 });