+++ src/config/react.js
@@ -4,18 +4,19 @@ const getOption = require('../lib/get-option');
 const includeRules = require('../lib/include-rules');
 
 module.exports = (configs, options) => ({
-	parserOptions: {
-		ecmaFeatures: {
-			jsx: true,
-		},
-	},
 	plugins: [
-		'filenames',
-		'promise',
-		'security',
+		'react',
+		'react-hooks',
+		'jsx-a11y',
 	],
 	env: {
 		browser: true,
+	},
+	parserOptions: {
+		sourceType: 'module',
+		ecmaFeatures: {
+			jsx: true,
+		},
 	},
 	settings: {
 		react: {
@@ -26,5 +27,7 @@ module.exports = (configs, options) => ({
 	},
 	rules: {
 		...includeRules('plugins/react', configs, options),
+		...includeRules('plugins/react-hooks', configs, options),
+		...includeRules('plugins/jsx-a11y', configs, options),
 	},
 });