+++ src/rules/ecmascript-2015.js
@@ -14,6 +14,7 @@ module.exports = (configs) => ({
 	'no-dupe-class-members': 'error',
 	'no-duplicate-imports': 'error',
 	'no-new-symbol': 'error',
+	'no-restricted-exports': 'off',
 	'no-restricted-imports': 'off',
 	'no-this-before-super': 'error',
 	'no-useless-computed-key': 'error',