+++ src/rules/strict-mode.js
@@ -1,9 +1,5 @@
 'use strict';
 
-const conditionalRule = require('../lib/conditional-rule');
-
-module.exports = (configs) => ({
-	strict: conditionalRule(configs, 'error', [
-		{configs: ['typescript'], value: 'off'},
-	]),
+module.exports = () => ({
+	strict: 'error',
 });