+++ src/rules/variables.js
@@ -1,21 +1,15 @@
 'use strict';
 
-const conditionalRule = require('../lib/conditional-rule');
-
-module.exports = (configs) => ({
+module.exports = () => ({
 	'init-declarations': 'off',
 	'no-delete-var': 'off',
 	'no-label-var': 'error',
 	'no-restricted-globals': 'off',
 	'no-shadow': 'error',
 	'no-shadow-restricted-names': 'error',
-	'no-undef': conditionalRule(configs, 'error', [
-		{configs: ['typescript'], value: 'off'},
-	]),
+	'no-undef': 'error',
 	'no-undef-init': 'error',
 	'no-undefined': 'off',
-	'no-unused-vars': conditionalRule(configs, 'error', [
-		{configs: ['typescript'], value: 'off'},
-	]),
+	'no-unused-vars': 'error',
 	'no-use-before-define': 'error',
 });